+++ QueueReceiverServiceTests/Services/PlantServiceTests.cs
@@ -1,0 +1,6 @@
+﻿namespace QueueReceiverServiceTests.Services
+{
+    public class PlantServiceTests
+    {
+    }
+}
